Added upload feedback features
Upload feedback is displayed with "Loading..." and the result string
returned by the ipc call. Clicks on the upload button before the ipc
call returns are ignored, as are clicks made without a selection.

Still missing asynchronous ipc calls; this is because sendSync in an
async function does not work as expected, and invoke is missing from
the ipcRenderer API.